Add support for physical pixel dimensions
This adds support for writing the pHYs (Physical pixel dimensions)
chunk to encoded PNGs (see https://www.w3.org/TR/PNG/#11pHYs).
The dimensions can be specified in either pixels per meter (which is
what the PNG format uses) or in dots per inch (which is commonly used
for images).