Only create the thread local Deflater on demand.
When getting a Deflater this has no performance implication. The JVM has
to do a null check on the deflater anyway before calling deflater.reset().

With this explicit null check (and the lazy initialisation) the JVM now
can prove that the deflater can not be null when calling reset() and can therefor
remove the null check on the reset() call.

This change saves a bit on native JNI resources (i.e. not used Deflaters, as usually
you are compressing all images with the same compression level).