Stream image data and extend the supported image data formats.
Also fix sub-image handling by correctly respecting the
scanline strides.

The inner row loops try to reduce as many calculations as possible
and also try to reduce the register pressure. In theory the JVM
should perform CSE and other SSA-form based optimisation on the
byte code when jitting the code, but in practice this
only works sometimes and is of course very JVM dependent.

By streaming the image data row by row we can get the data processed
by the predictor and the ZIP compressor while it's hot, i.e. in the
CPU caches.

If only one AbstractPNGLineConsumer is instanced in the live time
of the JVM it can inline the consumer or create a direct call. Otherwise,
it will fall back to a cheap vtable indirect call.

The asserts are only enabled when -ea is passed to the JVM. Which is the default in IDEA when running code. So ensure that any benchmarks are not running with the -ea parameter.