Handle premultiplied alpha when getting color values
PNG file should never contain color values with premultiplied alpha.
If the image we are saving has premultiplied alpha (determined by
the type), the color values of each pixel needs to be divided by
the normalized alpha value.

By adding this handling in getIntArgb() and get4ByteAbgr(), we can
remove the hack in toStream().
The added benchmark test shows that it is slightly faster do the
conversion pre-pixel instead of using Java to convert the type of
the whole BufferedImage before encoding.

Fixes #43